+++ aoe2/aoe2_settings.py
@@ -79,7 +79,7 @@ class AoE2OverlaySettings(RTSOverlaySettings):
         self.username: str = ''  # username
 
         # how to fetch match data: 'aoe2.net', 'aoe2insights.com' or '' for no match data
-        self.fetch_match_data = 'aoe2.net'
+        self.fetch_match_data = ''
 
         # layout
         self.layout = AoE2Layout()